-
The "Medical Premium Predictor" project employs machine learning algorithms to estimate medical insurance premiums based on a comprehensive set of individual attributes. By utilizing a Random Forest Regressor model trained on a diverse dataset encompassing factors such as age, height, weight, medical history, and more, this project offers a valuable tool for both individuals seeking insurance quotes and providers looking to offer accurate pricing.

This repository encompasses the codebase, dataset, and accompanying documentation for the project. It encompasses data preprocessing, exploratory data analysis, model training, and an interactive interface for seamless premium predictions. Furthermore, the project showcases insightful visualizations, shedding light on the intricate relationships between various features and premium prices.

Contributions and feedback are highly encouraged, as they play a crucial role in refining and enhancing the capabilities of this project in the realm of healthcare and insurance. Whether for personal use or as a resource for the industry, the Medical Premium Predictor aims to make informed insurance decisions more accessible.